requirements.txt
streamlit
python-docx
PyMuPDF
transformers
fuzzywuzzy
python-Levenshtein